Fixes
Found some things while releasing the book that I believe improve the
ClothMaker. The cloth folds better. There is now a menu item that allows
the distorted cloth to be copied to a Triangle Mesh. This make it
possible to use the result of the simulation in our scene or to export
the shaped cloth to another tool. I also added some randoness to the
fan.